* main: (1059 commits)
  Remove friendsofsymfony/jsrouting-bundle.
  Remove unused route exposing.
  Remove caldera_criticalmass_rest_track_ridelist.
  Remove caldera_criticalmass_rest_photo_ridelist.
  Remove caldera_criticalmass_rest_ride_show.
  Remove caldera_criticalmass_search_prefetch.
  Remove gregwar captcha bundle as we use friendlycaptcha.
  Remove laminas packages.
  Simplify xml loading as we are really just loading a wordpress feed.
  Improve form.
  Refactoring.
  Integrate to login form.
  Copy Captcha class from Wahlkampf.app
  Require friendlycaptcha.
  Remove unused gulp file.
  Disable xdebug.
  Delete this env file.
  Remove twitter config.
  Fix doctrine migrations config.
  These twitter handles do not exist anymore.
  ...

# Conflicts:
#	composer.json
#	composer.lock
#	config/routes/criticalmass/api.yml
#	src/Controller/Api/PhotoController.php
#	symfony.lock

+++ src/Controller/CalendarController.php
@@ -10,8 +10,8 @@ class CalendarController extends AbstractController
 {
     public function indexAction(Request $request, SeoPageInterface $seoPage): Response
     {
-        $year = $request->query->getInt('year', (new \DateTime())->format('Y'));
-        $month = $request->query->getInt('month', (new \DateTime())->format('m'));
+        $year = $request->query->getInt('year', (int) (new \DateTime())->format('Y'));
+        $month = $request->query->getInt('month', (int) (new \DateTime())->format('m'));
 
         try {
             $dateTimeSpec = sprintf('%d-%d-01', $year, $month);
